Add link_whiteboard field to Application model
This commit introduces the `link_whiteboard` field to the `Application` model as a new CharField with a max length of 255. The corresponding migration has been created, and the new field is integrated into the application details template for display and linking. This allows storing and viewing whiteboard links associated with applications.