Add group field to GitLabProject and sync with API namespace
This commit introduces a nullable `group` field to the `GitLabProject` and `HistoricalGitLabProject` models, linking them to `GitLabGroup`. It also updates the API parsing logic to map namespaces of type "group" to the corresponding `GitLabGroup` records in the database. These changes ensure proper association between projects and groups in both the database and API interactions.

Signed-off-by: iBfuRsiFUBbfzPdnJhXUfFFkk <jowly.knowing_2x@icloud.com>